Move AbSTT tests to AbTT
